Remove some un stable or not must features, make the library do what it should do. Add name to NSAssert message for better debuggability.
setupViewTree and updateViewFrame are 2 mainly used api, one to setup the view tree, which should be called in viewDidLoad or init. updateViewFrame do layout stuff and should be called in viewWillLayoutSubview or layoutSubView.

Note: Some api removed, it is not the time to take the burden of compatibility. Just make the lib easy to use.